Next experiment: custom attributes as member annotations
This next plug-in allows you to annotate struct members with attributes, and is capable of reporting the presence and arguments of those attributes.

The next logical step to actually achieving the goal here is to write pragmas that trigger generation of bitpacking code for specific struct instances. We'd want to generate bitpacking functions for nested structs as we encounter them (and ideally only if at least one instance of the struct is serialized in full, not split across a sector boundary).

We should also mark various attributes as mutually exclusive with one another.