String-related validation
Moved the code for parsing data members from struct.cpp to data_member.cpp, as prep for supporting union types.

We can now tell arrays-of-char apart from arrays-of-int8_t, even when typedefs are used.

The string-bitpack attribute now errors if used on something that isn't an array of [arrays of [...]] char.

Added missing pragma-onces.

TODO: Delete `typedef_map` as it's no longer needed.