notes and to-do list
running out of time to work on this little project for now. almost certainly won't get it finished in time. damn.

writing down the basic notes for it, then.

---

The recursive traversal function we use for sector generation now handles re-trying a field on the same go-around that serialization fails due to lack of space and inability to subdivide further. Made this change because the array case wasn't checking the return value of `_generate` and retrying, which was a mistake and would mean we'd lose fields; but if it did check and retry, then it'd be doing so in exactly the same manner as for struct members, leading to pointless code duplication; so why have the caller do it in both places when we can have the callee do it in just the one place?