work
More work on serialization.

Added functionality to the type-node wrapper: pretty-printing, and the ability to change the signedness of a type.

Added a deference operator to the value-node wrapper.

Added an iterator type to the list-node wrapper.

Wrote improved code (`bitpacking::stream_function_set`) for accessing and storing the pragma-specified bitpacking functions.