Started work on a rewrite of the lu-bitpack plug-in
Started work on a rewrite of the lu-bitpack plug-in, renaming the previous attempt to `lu-bitpack-failed`. I'll salvage code from it as appropriate.

For the rewrite, removed several "GCC helper" functions whose functionality already existed in the "GCC wrapper" code, and moved some more such functions into the wrappers. In particular, we now have wrappers for STRING_CST expressions, we can now retrieve the values of INTEGER_CST and STRING_CST nodes via the wrappers, we now have wrappers for TYPE_DECL nodes, and we can now walk a type's typedef chain using those and using accessors on `type`.