progress
We now have a way of passing the identifiers to serialize over to the sector functions generator.

The sector functions generator now generates a pair of top-level functions, which take a buffer and sector ID as arguments, and call the per-sector functions as appropriate based on the sector ID they receive.

Moved `sector_function_generator`'s recursive traversal out of a lambda and into a member function, to account for C++20 lacking deducing this (the alternative would've been a doubly-nested lambda that takes itself as an argument, to facilitate recursion).

Moved pragma handler functions to their own files.

Implemented heritable option sets.

Added a field attribute for opaque buffers.