fixes
The `lookup_name` function only retrieves DECL nodes, not TYPE nodes, and TYPE_DECLs only exist for uses of the `typedef` keyword. Added `type::lookup_by_name` to look up a type, preferring a TYPE_DECL if one is present but searching for a non-typedef type otherwise.

When creating a new FUNCTION_DECL, we now create the PARM_DECLs for the DECL_ARGUMENTS.

Various minor bugfixes.