fixes; we now generate top-level bitpack functions
When we create new function declarations, we now give names to the arguments, to avoid the need for as many null checks elsewhere.

The sector functions generator now reuses the declarations for the top-level functions, if they exist; and it now honors the requested top-level function names.

With these changes, we now successfully generate the top-level bitpack functions.