Testing; struggling to get generated functions to link
Added several accessors to the function, variable, and type wrappers.

Added an accessor to call expressions to try and get the callee.

I now have a makefile target (`test_and_run`) for compiling the test program and then running it. However, compilation is failing: the generated functions apparently aren't making it into the object files, so the linker can't see them.