Progress, but not enough
Figured out how to get the top-level serialization functions to emit to the object files, but I can't get the per-sector functions to emit the same way. It's worth noting that for the top-level functions, there are forward declarations in the code to be compiled, and we generate definitions; whereas the per-sector functions are generated from whole cloth without any presence in the source code.