The test program now builds!
I had forgotten to have the sector state create new root blocks when creating new functions, so we were continually recycling the same two root blocks (read and save) across all per-sector functions. That's now fixed, and the code for committing a function should now guard against it as well (by asserting that the local block you pass in isn't already in use).

With this change, and marking some more of the generated functions as non-extern, the test program now builds and runs!