Fixed auto-detection of string lengths
The serialization functions take the string's max length. When the string requires a null terminator, the max length is equal to the char array extent minus one.

When bitpacking options are specified for a string, indicating the presence of a null terminator but not specifying the length, then we need to deduce the length as the array extent minus one. We were previously always using the array extent (i.e. the behavior that's correct if the null terminator is optional in memory).