Opaque buffers verified to work
Renamed the `lu_bitpack_opaque_buffer` attribute to `lu_bitpack_as_opaque_buffer`.

We should now warn on any unhandled bitpacking attributes (in case we add an attribute definition, load it via the plug-in API, but forget to add a handler for it).

Added a test-case to verify that we're handling opaque buffers properly.