Major bugfixes; XML export
Changed the generated plug-in's internal name (in the makefile) from `lu-bitpack` to `lu_bitpack`. Putting a dash in a plug-in's name makes it impossible to read per-plug-in command line arguments, as dashes are used in the command line parameter name to separate the plug-in name from the argument key.

Implemented code to output XML describing the bitpacked format that we've computed and generated code for. You can specify the path to export XML to via the command line parameter `-fplugin-arg-lu_bitpack-xml-out=PATH/HERE.xml`. If you don't specify a (valid) path, we don't export any XML.

Fixed a bug that would cause us to index arrays out of bounds in some cases, when splitting them. When generating a for-loop to pack as many contiguous arrays as possible using a loop, we didn't clamp `can_fit` to avoid serializing past the end of the array.

Fixed the check for whether an indivisible field was too large to fit in a sector: forgot to convert the sector size from bytes to bits.

Fixed a bug that caused us to improperly compute the sizes of data values while packing into sectors, such that we were wasting space at the end of each sector. I had forgotten to check the bitpacking options in `codegen::member_descriptor::size_in_bits`.

Added `gw::decl::base::is_at_file_scope`.