Unified codegen and XML generation
XML generation no longer mimics the code generation process; rather, code generation now produces all of the output that we need for XML generation. (It occurred to me that the only thing we needed for XML was a list of what fields were written into what sectors, in what order; and that is something we can easily produce as part of codegen output.) This means we no longer need to maintain two separate processes and keep them perfectly in parallel.