Started work on a refactor of how we handle member bitpack options
My goal right now is to make it so that bitpack options can be set (via attributes) on types or on data members, with a data member applying the following options in order of increasing priority: heritable chosen by the member's type; attributes of the member's type; heritable chosen by the member; attributes of the member.

Not done yet. The `data_options` files are a bit too large, and functionality needs to be better separated; plus I haven't even attempted to compile any of it yet.