Bitpacking attributes on types now work
Created a testcase for bitpacking attributes on types.

Modified the attribute definitions so that they can be applied to types, typedefs, and field declarations.

We now handle attributes applied to types more reliably.

Modified the `debug_dump_identifier` pragma so that when it's used on a TYPE_DECL, we also dump the *_TYPE node.