Work and plans for attribute handlers
Moved documentation of attribute handlers out of the `no_op` handler and into a Markdown file.

Added custom attribute handlers for the bitcount and range attributes.

---

I'm strongly considering moving error reporting for bitpack attributes out of the codegen system and into attribute handlers. See the to-do list for details.