Moved most baseline attribute validation to attribute handlers
Implemented the "lightswitch" pragma.

Moved basic validation of string attributes to the attribute handler. We store the validated length and with-terminator flag in internal-only attributes.

Got rid of the `processed_bitpack_attributes` class. The new approach is to set an "invalid" attribute if processing of any attribute fails, and only allow valid attributes to be retained.

Fixed `string_constant::value_view`; I was returning `std::string` by mistake (bad copying and pasting), so we were returning a view to a soon-to-be-destroyed implicit local.

Still need to do baseline validation and formatting of the pre-pack/post-unpack attribute.