"Inherit" attribute now does more validation
The `lu_bitpack_inherit` attribute handler now validates whether the specified heritable exists, and errors if used more than once on the same entity.

The `bp_attr_context` helper class now makes it easy to check whether the current attribute has already been applied directly to the to-be-annotated entity, or whehter the attribute tried and failed to apply.