We now use GCC attr nonstring instead of custom stuff
The `lu_bitpack_string` attribute no longer takes any arguments. The max length should always be computed from the data member or from a heritable, and the need for a string terminator is indicated by the absence of GCC's built-in "nonstring" attribute.

It's worth noting that "nonstring" is only allowed on non-type DECLs, not on types or typedefs. It would only be useful on typedefs if bitpacking options in general worked on typedefs, of course.

When you specify a default value for a string, we now verify that the default value fits in the string.

When we generate code to write a default value to an omitted-and-defaulted string, we now check if the default string is smaller than the destination array. If so, we generate a memset call to zero out the rest of the array.

String constants can no longer be created from an `std::string_view`. They're supposed to include a null terminator, but we can't guarantee that one is present when dealing with `std::string_view`. Currently we require a `std::string` or `const char*` argument; we could change this to use a `lu::strings::zview` instead.