Buffer array codegen; started work on transforms codegen
Arrays of opaque buffers now serialize with a single buffer-read/buffer-write call, instead of looped calls.

Computed transform options now store the transformed type, for easier retrieval during codegen.

Added `serialization_value_path::transform_to` as prep for logging transformed fields during codegen, for the XML output.

"Transformed" is now a `member_kind` with incomplete accessors and queries on `serialization_value` and friends.

Added an incomplete `_serialize_transformed` function to `sector_functions_generator`. This is only appropriate for when we know an entire transformed value will fit; we still need code to deal with splitting transformed values across sector boundaries when possible and pushing them to the next sector when not.

Wrote a to-do list for finishing the transforms implementation.

Removed information on heritables from the readme.