Tests and fixes
Created testcases for arrays of opaque buffers, and for typedefs used to apply options to array types.

Fixed an issue in `bp_attr_context` that caused it to choke when attributes were applied to typedefs. It's still not perfect, though.

Fixed a mistake in `data_options::requested` that caused it to fail at the end of a transitive typedef chain, by trying to act on a non-existent typedef for the last type.