Removed "string_char" global option; other fixes
Removed the ability to specify a string character type as a global option. We now only validate that the argument type to the string bitstream functions is a pointer to single-byte integer values.

Since global options are loaded at the time they're defined (rather than at codegen time), we have to remember whether they were invalid so that codegen can fail if they were. Doing this fixes bugs where codegen would crash because it'd blindly pull stuff out of the global options trusting that it'd already been checked for correctness.