Plans for a rewrite
I don't think it's possible to implement transformations within the current system -- or at the very least, it's overwhelmingly difficult, and I've only had false start after false start whilst trying. I think I need to do a total rewrite, going back to more-or-less how things worked when I was doing XML-based proofs of concept: we need to generate a list of "serialization items" reflecting the things to bitpack, and then recursively expand and split them as necessary to fit them into sectors.

The benefit of this approach is that in addition to simplifying the logic for transforms, it would also make tagged unions very, very, very easy to implement as well.