Update TODO.md
Planning out precisely how codegen should work. The flat serialization item list is perfect for allotting values to sectors, but it's not so good for codegen, where we need nested code blocks: during codegen, we need to know which blocks to write expressions to, and we need to know when to open or close potentially nested blocks.

Easier to have a concept of "instructions" that form a tree. We go from serialization items to instructions, and then from instructions to code.