Redesign plans
After some prototyping in JavaScript, trying to work out how to take a flat list of serialization items and produce an instruction tree that would map to cleaner codegen, I've come to the conclusion that the current serialization item format is inadequate, failing to handle the case of unions nested inside of transformed structs cleanly.

I've designed a new path format, and intend to prototype the logistics in JavaScript before proceeding further. The C++ `serialization_item` class would not require terribly significant changes, but I want to make sure I've got the logic in place for list-to-tree before I do any rewriting on the C++ side of things.