JavaScript prototyping for serialization items
Converting item lists to instruction trees is, so far, a two-stage process. First, we "re-chunk" the list items, splitting them at any point where we'd generate parent/child nodes: conditions and transformations. The result is still a flat list, but the items (the paths) are split differently. Then, we traverse that list to generate a final node tree.

Re-chunking works. Converting the re-chunked items to trees, however, is broken.