JS prototyping almost done, but more design changes needed
In the JavaScript prototype, we can now convert a list of serialization items through a list of re-chunked items into a tree of instructions. I've written some lengthy code comments which should help walk me through porting this to C++...

There are, however, some holes even in this design. The to-do list has specific explanations and instructions on remedying the issues, and notes therein should be considered the most up-to-date, overriding anything else in this commit.