JavaScript prototyping is almost done
The JS prototype can now parse very simple C attributes. The `transform_to` attribute is used to indicate a transformed type, and is supported on types or fields.

All parsed C-types are now tracked in a central dictionary, so that we can handle transitive transforms.

Revised how we generate expanded serialization items, so that when a value is transformed, we access members of the transformed type. Note that the JavaScript prototype doesn't understand typedefs, so transformations to arrays aren't represented here.

Revised how conditions within re-chunked items are stored, so that we can properly resolve loop indices therein when we generate a node tree.

Refactored the prototype's HTML and CSS to simplify the latter.

At this stage, we can properly generate an instruction tree from a list of re-chunked items, accounting for transforms, conditions, arrays, and any arbitrary nesting of those three.