Returning from JavaScript to C++, things mostly work
Ported the JavaScript prototype code over to C++, with some additional changes to enable support for zero-padding. Initial tests seem to work, though the output is not entirely as we expect because the sector-splitting code mishandles conditions (a problem that would've been present before the current redesign of serialization items).

Fixed an edge-case in sector splitting that could cause us to try and pack elements into a sector with zero bits remaining. In the case of serialization items that can be split (i.e. padding), this would result in us splitting an item into two items, one of which was zero-size.

Sector splitting no longer force-expands unions during the initial scan. Instead, expansion of unions, anonymous structs, and arrays thereof is done as a final post-process step, to ensure that it doesn't interfere with the post-process step to convert sequential array elemetns to array slices.