Re-chunked items now make arrays explicit; documentation updates
When we generate re-chunked items, we now fully expand arrays so that array slices are explicit. That is, given a field `int foo[4][3][2]`, a serialization item for `foo` will now produce the same re-chunked item as a serialization item for `foo[0:4][0:3][0:2]`.

Updated the documentation for serialization items. Documented re-chunked items and instruction nodes.