Documentation and to-do list updates
Updated the documentation on `array_slice` nodes to note the count-1 case, to talk about how I've set things up for us to handle it, and to note recommendations for guarding against the possibility of that setup failing.

Updated the to-do list.

Came up with an idea for completely eliminating the count-1 edge-case. I can tackle it tomorrow, if there's time.