Array slice improvements; debug dump improvements
Instruction tree generation: simplified the code for handling condition operands, by allowing `_current_value_for_stack` to refer to just a slice of a stack if need be.

Instruction tree generation: implemented the planned approach for avoiding single-element array slices in favor of direct references to the single array elements. In tests, it works!

When we debug-dump re-chunked items, we now indicate whether they're omitted and defaulted.