Basic codegen works; unions crash the compiler
Fixed a mistake in `value_path` that caused production of a `value_pair` to fail: the first segment is never member access; it is always direct access to a DECL.

Fixed a mistake in `value_path` that caused us to use the wrong variable when producing a `value_pair`.

Fixed code generation for `array_slice` instruction nodes: we now force the generated loops to use our pre-created loop counter variables.

Rebuilt the `generate_functions` pragma so that it's no longer a no-op.

Code generation works for non-union values, but the compiler internals choke on the code we're generating for unions.