Bitstream fixes; whole-struct function fixes
Fixed how my bitstream functions handle reading signed integers from the bitstream.

Updated `decl_descriptor` so that if we generate a descriptor for a pointer-type PARM_DECL, we unwrap the pointer type. PARM_DECL descriptors are only used for whole-struct functions, where the struct is always passed as a pointer and the struct is what we wish to (de)serialize, so this change is necessary so that we can properly query the members of those descriptors.

Created and fixed a test-case with a top-level non-struct VAR_DECL.

When we generate instructions to serialize a `char`, we now use the bitstream function for `uint8_t`. Previously, we had no case for this, so we'd fail to select bitstream functions, fail an assertion, and crash.

Code generation should now halt if any to-be-serialized decl is detected as having errors.

Fixed the attribute handlers so that they no longer reject VAR_DECLs.