Fixed handling of string-type default values in codegen
Pretty-printing function types should now include the varargs ellipsis, if one is needed.

Pretty-printing parameter types now includes the "restrict" qualifier.

Improved how the type wrapper handles type qualifiers. Added the ability to query and alter the "atomic" and "restrict" qualifiers. Added the ability to alter the "volatile" qualifier.

Added `gw::type::base::has_c_boolean_semantics`. In versions of GCC that support setting underlying types on C enums, this function will return true for booleans or for enums whose underlying types are booleans.

Added `gw::type::base::is_bitint`. In versions of GCC that support _BitInt types, this returns true for them.

Added an experimental function `gw::type::base::is_assignable_to`. Untested, as I realized after writing (porting?) it that I don't need it yet. Marked as untested in the source code.

Changed the behavior of `gw::type::is_arithmetic` to include a broader range of integral-but-not-integer types.

Modified `gw::type::function::for_each_argument_type`: it now invokes the passed-in functor on a `gw::type::base` rather than a raw tree node, no longer invokes the functor on the no-varargs sentinel, and is capable of passing the current argument's (zero-based) index to the functor if the functor opts to accept it.

Code generation is now more reliably able to access the `memcpy` and `memset` functions, grabbing them ahead of time and validating their signatures against what we expect. This means that `codegen::instructions::single` doesn't have to grab them every time -- and doesn't risk failing to, either.

If bitpacking attributes are seen before codegen is enabled, we now remember the location of the first seen attribute. if codegen is attempted, we print that location.

The `bitpacking::data_options::computed` structure now stores whether an annotated DECL had the "nonstring" or "lu_nonstring" attributes. This makes it far easier to check for those attributes down the line, particularly when generating code for omitted-and-defaulted fields' string-type default values.