Error-handling fixes
We now pre-check all to-be-serialized values (including any nested fields) before attempting codegen, to ensure that errors on nested structures don't cause us to choke.

Fixed error-checking on requested options so that we reject unions without tags, but only if they're not opaque buffers, and so that we don't repeatedly emit the same error messages. (Errors are checked for per-type, but then we have to re-check per DECL of those types.)

Wrote a GCC wrapper for the PLUGIN_TYPE_FINISHED callback so that we can prevent it from re-firing on the same types repeatedly.