Started work on XML output
Started work on the XML output.

We diverge from the previous attempt in that the XML output is based on the codegen instruction nodes rather than on the serialization items. Conditions make dealing with the serialization items... inelegant... for this use case; and in any case, having the most direct representation of the generated code possible will reduce the likelihood of bugs leading to XML output that is inconsistent with the generated code.

Made `walk_instruction_nodes` usable on const node references.

Identified a major flaw in how we handle sector splitting with respect to union-related conditions.