Stats reporting
We now gather and emit stats about bitpacked types, in the XML report. Currently we emit the max sector count and size; for each sector, we emit the total bits used; and for each type, we emit the total count seen per sector (structs split across sectors only count for the first sector), the C sizeof and alignment, and the total packed and unpacked sizes in bits.