Started work on splitting the GCC wrappers (WIP)
Started work on splitting the GCC wrappers into reference-like and pointer-like wrappers. I'm doing this in the `gcc-wrapper-redesign` folder.

Updated the old wrapper code with some small fixes.

Wrote documentation for the new set of wrappers.

Updated the to-do list. I'm thinking of renaming "reference-like" and "pointer-like" wrappers (`node` and `node_ptr`) to "valued" and "nullable" wrappers (`node` and `node_nb`). This would eliminate some of the connotations of the word "reference" (e.g. no rebinding), and it would allow us to ditch `operator->` in favor of just having `operator*` akin to `std::optional` (which is good because the former requires an icky pointer cast due to how the wrappers work).