Redesigning GCC wrappers again
Moving from "reference-like" and "pointer-like" to general and optional.

Added a wrapper for identifier nodes, though it's not used in many places right now. The goal is to get rid of as many direct uses of <stringpool.h> and `get_identifier` across the codebase as possible.

The `list_node` wrapper is now a `node` subclass again, instead of being its own separate thing.

Made `list_node::prepend` safer to use in a variety of scenarios, including when the list head is referenced by something which is meant to own the list as a whole.

Renamed `node::as_raw` to `node::unwrap` for consistency with `optional_node::unwrap`.

Several accessors that returned `T_ptr` no longer manually wrap `tree` return values in `optional_T`, instead returning the `tree` pointer directly and relying on implicit conversion/construction.