GCC wrapper rewrite: plug-in builds but fails to run
Fixed a slew of compiler errors.

Moved accessors for a variable's initial value from `decl::variable` to `decl::base_value`, since they should apply to other value DECLs (e.g. PARM_DECL).

Fixed the node wrapper boilerplate to ensure that we can always `wrap` a node with any wrapper of the appropriate type.

Added code for a test plug-in that just debug-prints info about identifiers, using the wrappers.

The GCC wrapper rewrite test plug-in now builds. However, the testcase fails on a typecheck when trying to retrieve the scope (DECL_CONTEXT) of a TYPE_DECL.