GCC wrapper rewrite: critical fixes
Improved error reporting for when we fail to wrap a tree in a wrapper, or when we fail to cast a wrapper.

Fixed a mistake in `optional` wherein we'd always fail when trying to wrap NULL_TREE, because we did the type checks unconditionally.

Moved "initial value" accessors from `decl::base_value` to `decl::variable`. For function parameters, the initial value is stored on the function type. (The function type's arguments are encoded as key/value pairs, where the keys are the initial value EXPR/CSTs and the values are the argument TYPEs.)

With these fixes, the test plug-in now works and can properly produce debug output.