Ported over attribute handlers and bitpacking namespace
Ported over attribute handlers and most of the `bitpacking` namespace.

Redesigned data options in the process of porting them over. They needed to be ported now because attribute handlers for unions need to be able to parse out data options in order to compare the leading members-of-members of internally-tagged unions. As part of this redesign, we no longer have the `member_kind` enum; instead, we take advantage of the "X-options" variant being... well, a variant -- a tagged union. We can check `data_options::is` and use `data_options::as` as accessors.

Updated the constructor for `gw::list_node` to hopefully make it a bit less of a footgun. The class design is better for reading lists that already exist than for creating new ones... I've written documentation explaining the reasons for this design.

`gw::identifier::name` now returns a `lu::strings::zview` rather than a `std::string_view`, so that it doesn't have to be unwrapped (to `const char*`) and rewrapped in order to use it in wrapper APIs that require a Z-view (i.e. guaranteed-null-terminated view).

Updated the to-do list.