Codegen: tests and fixes for opaque buffers and strings
Fixed a codegen bug wherein we'd try to split serialization items that represented opaque buffers, if they referred to anonymous structs or to unions.

Fixed a mistake made when refactoring data options, that caused us to fail to handle the `nonstring` attribute (and `lu_nonstring`, our custom counterpart for types) at the last moment.