Transform fixes; max sector count fix
Fixed a bug with transitive transforms split across sectors, wherein we were forgetting to generate the first "pack me to unpack me" call.

Fixed a bug with decl descriptors wherein we generated the list of transformed types incorrectly, such that our serialized type was not the final transformed type as intended.

We now actually enforce the maximum sector count.