Ported over the XML output.
Modified `codegen::stats::serializable` to always have total bitcounts. They were only optional because in the previous plug-in draft, the same fields were present for things that didn't always use them (e.g. sectors).

Ported over the codegen stats gatherer. Some changes have been made: we no longer use a state variable to track what sector we're in, nor use a separate recursive function to count how many of each stat-tracked entity appears inside of each top-level identifier. Rather, we use a "context" object ferried as an argument to the calls that need it, and we perform all stat-gathering on the serialization item list (reasoning that the first segment in any non-padding item should be the top-level identifier containing the to-be-serialized value).

Ported over the code to generate XML reports. It's a bit spaghetti, but better than it was last time.