GCC wrapper fixes; codegen now generates builtins
Fixed the GCC wrapper changes from the last commit.

Added a serialization item list op to get the total serialized size of a list, without having to build a vector of pairs. Should update the stat-tracking for sectors to use this too.

Code generation now also generates two file-scoped (and constexpr in C23) variables with static linkage: `__lu_bitpack_sector_count` and `__lu_bitpack_max_sector_size`. These are defined to equal the settings you passed in for sector generation or, if you generated only a single sector with no maximum size, the latter will indicate the final size of the sector (so you know how much memory to allocate for the bitstream).