Fix for building for GCC 11.5.0
Fixed version checks that broke for GCC 11.5.0. Apparently, the date on which a commit is made doesn't line up cleanly with the release dates for GCC versions. Haven't been able to find how much lag time there is, though, so given the date of a commit in the GCC repos, I still don't know how to tell when that commit actually shipped to users.

Updated gitignore to try and avoid committing build output for testcases (i.e. the test programs).