Update setup.sh
Factored the M4 check into a variable to match the PV check.

Some of the checks we use for fast builds are inconsistent: either $fast is testing inconsistently, or a directory-exists check we run is testing inconsistently. Not sure why yet. For now, I've added an else branch that detects when the two checks produce inconsistent results, with a special message written for that case. Guess we'll see if that triggers.

Also -- current plug-in source code tests and seems to work well enough with GCC 12.4.0.