GCC install script improvements
Fixed buggy checks for installed programs.

We now set the terminal window title throughout the installation process. Several steps of the process produce mountains of log output, so setting the window title makes it possible to know where we are in the overall process even when we can't scroll far up enough to see my messages.

Still investigating why we're failing to set up GMP with fast builds enabled. Added a 0.2-second delay on GMP folder checks and am testing that now; I"m wondering if it's some kind of timing problem.