ugh
Documentation for GCC wrappers: moved the changes needed for C++ support to their own document.

GCC install script: originally, we'd detect if you had the archive downloaded and skip re-downloading it if so. However, this fails if the download was interrupted. As such, we now re-try the download if we skip downloading but can't unpack what's already there.

GCC install script: fixed checks for installed programs (M4 and PV). Again. I love bash scripting: the fact that the thirty syntaxes it has for everything only barely qualify as "deterministic" adds fun and excitement to scripting. It's like a gacha game, except that what you're gambling with is your limited, precious time on this earth.