GCC 13+ and more GCC install script edits
Tested the plug-in in GCC 13.1.0 and 13.3.0 using the `codegen-various-a` testcase. No compile errors. Run-time test output looks correct.

---

Updates to the GCC install script: you can now specify the thread count to use, with it defaulting to nproc, rather than it being locked at nproc. This can be used to install multiple GCC versions (via multiple invocations of the script) simultaneously, since several steps of the overall installation process don't parallelize very well.

Made another attempt to refine the GMP directory check, though I'm not in a position to check that anymore because it looks like GCC 13+ makes sure GMP is available even after a non-bootstrapping build, rendering the check pointless for those versions.

The script is getting bizarre parse errors where bash is choking on tokens that are inside of line comments (per displayed excerpts), or claiming that the letter "c" is an unrecognized command (with obviously wrong line numbers). Not sure how to fix that. Maybe I should just say "screw it," figure out how to recreate the script in Lua, and have the bash script just install Lua. Honestly, that'd probably be a far better option than dealing with an interpreter that is, somehow, even more broken than CMD.EXE.