14.2.0 fix; redid the GCC install script in Lua
`gw::value`: Turns out, the `build_component_ref` change that adds a bool parameter hasn't even shipped yet. Going to make a blind guess that it'll ship in 14.3.0. Updated the version check accordingly so this should no longer fail to compile in 14.2.0.

GCC installation script: I'm no longer going to maintain the bash script, and am instead using a Lua 5.4 script. Having a script interpreter that's actually deterministic, with a consistent and reliable syntax, feels downright luxurious after the last day or so of work.