Fix to gw::scope; improved debug logging
GCC wrappers: Fix for a crash when retrieving the scope of an enumeration constant DECL.

Plug-in: enhanced the `debug_dump_identifier` pragma to require different operators for accessing members of types versus values, and to allow member access in the first place. Backported these enhancements to `debug_dump_bp_data_options`.

Added a testcase named `constexpr`, and edited the makefile so that this testcase specifically compiles as C23.