Fixed code generation for `int`-type integers
If an integer value's canonical type is `int`, we now use the u8, u16, or u32 functions based on the size of `int` on the current platform. We still have no support for 64-bit integers at this time. This fixes an assertion failure when doing code generation for bare `int`s on some architectures and compilers.