It's now possible to serialize the data a pointer points to
When requesting codegen, you can now write `*identifier`, `**identifier`, etc., to request that the generated code serialize not the pointer itself, but the data to which it points.

To facilitate this functionality, `decl_descriptor` instances that wrap VAR_DECL or PARM_DECL now store a "dereference count," indicating the number of times the DECL should be dereferenced when generating serialization code. The count also affects how we handle the data type; in short, it influences what we're actually describing.

Previously, we had a nasty hack for whole-struct functions, which access a to-be-serialized struct via a pointer argument: `decl_descriptor` instances for PARM_DECL were assumed to be those arguments and dereferenced the PARM_DECL one time; and `value_path` made a similar assumption and dereferenced blindly. This hack has been removed; we now use the new dereferencing feature when creating descriptors for these function arguments.