Various fixes
"Single" instruction nodes now use range checks on the bitcounts, to account for bitfields (which generate bespoke types; we're not operating on the FIELD_DECL so it's hard to find the original declared type).

The constexpr constants we generate are now `const size_t`; additionally, the variables themselves are marked as constants (i.e. "read-only"), not just their types.