Critical fix to bitpacking of pointers, and to sector splitting
Fixed how we compute the serialized size of bitpacked values. We were failing to handle pointers and transformed values (when the transformed type was not a struct or union).

Fixed how we handle sector splitting when multiple distinct identifier groups are requested (e.g. `a b | c`): we were accidentally discarding all sectors save for those in the last identifier group.

Fixed how we parse identifier dereferencing in the `generate_functions` pragma.