Critical fixes to the "where did it end up" pragmas
Fixed several edge-cases in the pragmas that allow you to know where a given value ended up within the bitstream. The code for this is hideous and not terribly efficient, but it works.

If you ask about `foo` and the bitstream contains `foo[0]` through `foo[5]`, we need to report information about `foo[0]`; yet if you ask about `foo[3]` and the bitstream only contains `foo` as a whole unit, then we need to compute where `foo[3]` would be. The easiest way to do this, and the way that assures consistency with codegen, is to just have `last_generation_result` store a fully-force-expanded copy of all of the serialization items per sector... so that's what we do. It runs surprisingly fast despite the monstrous amount of heap allocation it must be doing.