XML reports now sort categories and types
The XML report generator now sorts categories and types deterministically before outputting them as XML. As with the previous commit, this should reduce the number of spurious changes in the XML output after successive codegen runs, which is good for diffs and commit histories.

Categories and types are sorted more-or-less by codepoint (which maps to alphabetical sorting in English). For types, there are some nuances since a single type can have multiple identifiers (look up C name spaces, not to be confused with C++ namespaces).